Think I'm finally happy with HardwareRMT. It now uses a single sync instead of start and stop

+++ Code/C++/Shared/Communication/Interfaces/IChannel.hpp
@@ -8,8 +8,8 @@ public:
     virtual ~IChannel() = default;
 
     virtual bool configure() = 0;
-    virtual int readBytes(char* dst, size_t maxlen, uint32_t timeoutMs) = 0;
-    virtual int writeBytes(const char* src, size_t len) = 0;
+    virtual size_t readBytes(char* dst, size_t maxlen, uint32_t timeoutMs) = 0;
+    virtual size_t writeBytes(const char* src, size_t len) = 0;
     virtual void flushTx() = 0;
     virtual const char* getDescription() const = 0;
 };
