+++ Code/C++/Shared/Communication/UART/HardwareUART.hpp
@@ -15,7 +15,7 @@ public:
     bool configure(
 	) override;
 	virtual ~HardwareUART() override;
-    int readBytes(char* dst, size_t maxlen, uint32_t timeoutMs) override;
-    int writeBytes(const char* src, size_t len) override;
+    size_t readBytes(char* dst, size_t maxlen, uint32_t timeoutMs) override;
+    size_t writeBytes(const char* src, size_t len) override;
     void flushTx() override;
 };
